Moved sqlite.py into a data_sources module.
